src/main/java/io/lumigo/core/parsers/SnsParser.java
package io.lumigo.core.parsers;

import com.amazonaws.Request;
import com.amazonaws.Response;
import com.fasterxml.jackson.databind.JsonNode;
import io.lumigo.core.utils.JsonUtils;
import io.lumigo.models.HttpSpan;
import java.util.List;
import org.pmw.tinylog.Logger;

public class SnsParser implements IAwsParser {
    @Override
    public void parse(HttpSpan span, Request request, Response response) {
        String topicArn = getParameter(request, "TopicArn");
        span.getInfo().setResourceName(topicArn);
        span.getInfo().setTargetArn(topicArn);
        span.getInfo().setMessageId(extractMessageId(response.getAwsResponse()));
    }

    private String extractMessageId(Object response) {
        try {
            JsonNode node = JsonUtils.convertStringToJson(response.toString());
            return node.get("messageId").asText();
        } catch (Exception e) {
            Logger.error(e, "Failed to extract messageId for SNS response");
            return null;
        }
    }

    private String getParameter(Request request, String key) {

        if (request.getParameters() != null
                && request.getParameters().get(key) != null
                && ((List) request.getParameters().get(key)).size() > 0) {
            return ((List) request.getParameters().get(key)).get(0).toString();
        }
        return null;
    }
}
